integrate constructor tear-off tests
