+++ test_data/rules/always_specify_types.dart
@@ -4,7 +4,17 @@
 
 // test w/ `dart test -N always_specify_types`
 
+// ignore_for_file: unused_local_variable
+
 import 'package:meta/meta.dart';
+
+/// Constructor tear-offs
+void constructorTearOffs() {
+  List<List>.filled; // LINT
+  List<E> Function<E>(int, E) filledList = List.filled; // OK - generic function
+  filledList<int>(3, 3); // OK
+  filledList(3, 3); // OK - generic function invocations are uncovered -- see: #2914
+}
 
 typedef MapList = List<StringMap>; //LINT
 typedef JsonMap = Map<String, dynamic>; //OK
@@ -54,6 +64,7 @@ main() {
     print(i);
   }
   List<String> ls = <String>[];
+  // ignore: avoid_function_literals_in_foreach_calls
   ls.forEach((s) => print(s)); //LINT [15:1]
   for (var l in ls) { //LINT [8:3]
     print(l);
@@ -68,13 +79,14 @@ main() {
     print(e);
   }
 
+  // ignore: non_constant_identifier_names
   var __; // LINT
 
   listen((_) { // OK!
     // ...
   });
 
-  P p = new P(); //OK (optionalTypeArgs)
+  P p = P(); //OK (optionalTypeArgs)
 }
 
 P doSomething(P p) //OK (optionalTypeArgs)
@@ -82,7 +94,7 @@ P doSomething(P p) //OK (optionalTypeArgs)
   return p;
 }
 
-listen(void onData(Object event)) {}
+listen(void Function(Object event) onData) {}
 
 var z; //LINT
 
@@ -96,8 +108,8 @@ class Foo {
 }
 
 void m() {
-  if ('' is Map) //OK {
+  if ('' is Map) //OK
   {
-     print("won't happen");
+    print("won't happen");
   }
 }
