--- test_data/rules/experiments/constructor_tearoffs/analysis_options.yaml
@@ -1,3 +1,0 @@
-analyzer:
-  enable-experiment:
-    - constructor-tearoffs
