--- test_data/rules/experiments/constructor_tearoffs/rules/.mock_packages
@@ -1,1 +1,0 @@
-meta:../../../../mock_packages/meta/lib/
