+++ test_data/rules/unnecessary_lambdas.dart
@@ -9,6 +9,32 @@
 // ignore_for_file: prefer_expression_function_bodies
 // ignore_for_file: unused_local_variable
 
+typedef Maker = Object Function();
+
+class C {
+  int c;
+  C([this.c = 3, int y]);
+}
+
+var l = <Maker>[
+  C.new, // OK
+  () => C(), // LINT
+  () => C(3), // OK
+];
+
+void constructorTearOffs() {
+  var listOfInts = [1, 2, 3];
+  for (var c in listOfInts.map((x) => C(x))) { // LINT
+    print(c.c);
+  }
+  for (var c in listOfInts.map((x) => C(3))) { // OK
+    print(c.c);
+  }
+  for (var c in listOfInts.map((x) => C(x, 3))) { // OK
+    print(c.c);
+  }
+}
+
 int count(String s) => s.length;
 final strings = ['a', 'bb', 'ccc', 'dddd'];
