+++ test_data/rules/unnecessary_statements.dart
@@ -4,6 +4,15 @@
 
 // test w/ `dart test -N unnecessary_statements`
 
+void constructorTearOffs() {
+  MyClass.new; // LINT
+  MyClass.named; // LINT
+
+  var m = MyClass.new;
+  m().foo; // LINT
+  m().field; // LINT
+}
+
 String f(Object o) {
   // See: https://github.com/dart-lang/linter/issues/2163
   o as int; // OK
@@ -172,8 +181,11 @@ get getter => true;
 int field;
 
 class MyClass {
-  int field;
+  int field = 0;
   bool foo() => true;
 
   get getter => true;
+
+  MyClass();
+  MyClass.named();
 }
